新增 Customer Controller Insert 新增  Logic CustomerLogic InsertCustomers 修改 CustomerViewModel email 型別 修改 CustomersRepository 修改 CustomerModel age 型別

+++ InterviewDemo/Models/CustomerModel.cs
@@ -10,7 +10,7 @@ namespace InterviewDemo.Models
         public int id { get; set; }
         public string firstname { get; set; }
         public string lastname { get; set; }
-        public ushort age { get; set; }
+        public int age { get; set; }
         public DateTime birthday { get; set; }
         public string email { get; set; }
         public DateTime createdate { get; set; }
